A lot of bugfixes for compression workbench as well as runtime improvements. Implementation fo a HashNode which tries to increase the runtime, is however slower as the leaders node. Further the adoption of the hierarchy import manager to account for brushing was attempted, some error still exists.

+++ PCViewer/compression/HierarchyCreateNode.hpp
@@ -26,7 +26,7 @@
 class HierarchyCreateNode{
 public:
     virtual void addDataPoint(const std::vector<float>& d) = 0;
-    virtual HierarchyCreateNode* getCacheNode(int& cacheScore) = 0;
+    virtual HierarchyCreateNode* getCacheNode(long& cacheScore) = 0;
     virtual void cacheNode(const std::string_view& cachePath, const std::string& parentId, float* parentCenter, float parentEps, HierarchyCreateNode* chacheNode) = 0;
     virtual size_t getByteSize() = 0;
     virtual std::shared_mutex& getMutex() = 0;
